Removed abstraction and refocus on "new" model only.
For the sake of simplicity @MaxVinValk  and I decided to remove the abstraction and to only implement a single model.

Includes some changes to the drop-down menu, including removal of message option (not required here) and addition of network wide dissonance updates (can simulate external events but currently just there for fun).

Also includes some updates to edge coloring and the node coloring. For the latter now dissonance levels are taken into account (e.g. too high dissonance + belief > .5 == pink).

Further I removed the message class and the prunedissidents method on node level since it is implemented in the receiveMessage step.

JK