Included test data and adjusted .gitignore
Adjusted .gitignore to keep the distribution sampler local for now until I receive more information on what we can include in the public repository.

Added test data sampled from a multivariate normal with means = 0 and covariance matrix being an identity matrix (3x3). Each row in this file should be used to initialize the parameters for an agent.

@MaxVinValk  you'll notice that there are three parameters/columns in here but currently only two relevant parameters are implemented in the model. I will work on implementing the final one over the weekend so keep this in mind for your testing. :)