First version including extraversion + sampler to initialize parameters based on real data
Included a few rules to include the extraversion parameter, which influences the personal network size of each individual agent (soft connection limit). Additionally, the parameter currently influences the effectiveness of the dissonance boost when a new connection is made or a conflicting source is removed from the agent's network.

Adapted the consideration of another agent's belief: openness is weighted using the ratio of current dissonance / dissonance treshold. Individuals that experience more dissonance therefore become more narrowminded.

Also included a sampler class written in Python. This one is used to initialize the agents based on real data. The sampler uses a KDE in case a dataset is provided to estimate the PDF for the data at hand before drawing samples. If no sample data is provided it uses a multivariate normal approximation using the mean and covariance structure obtained from the dataset. The datset uploaded here is NOT the original dataset but contains values that were re-sampled based on the PDF estimate of the original dataset.

Bugs:
The dissonance updating function does not work well with the imposed range of 0-1. This means that when using the real dataset and the polarizing technique a tremendous amount of individuals just disconnects from the network and never attempts to re-connect. We need to evaluate this and discuss alternatives @MaxVinValk  and @TomCTheuer