Fixed isolating behavior of nodes + UI elements tto control weights
There was a problem with the calculation of dissonance that became evident only when using parameters inspired by real data. Dissonance is now calculated correctly at every encounter. We still need to look into the limits of the function.

Finally, there was an issue with the recommendation strategy - it did not take into account the new calculation of connection limits and would therefore suggest agents even though they would never accept a new connection. This contributed a lot to the isolation we were observing.

Added two UI elements allowing to change the weighting of the dissonance ratio which penalizes openness and openness itself.